# frontend performance: Recent topics to stream renarrows.