--- a/static/js/recent_topics_ui.js
+++ b/static/js/recent_topics_ui.js
@@ -17,7 +17,6 @@ import * as message_view_header from "./message_view_header";
 import * as muted_topics from "./muted_topics";
 import * as narrow from "./narrow";
 import * as narrow_state from "./narrow_state";
-import * as navbar_alerts from "./navbar_alerts";
 import * as navigate from "./navigate";
 import * as people from "./people";
 import * as recent_senders from "./recent_senders";
@@ -684,9 +683,8 @@ export function hide() {
     // before it completely re-rerenders.
     message_view_header.render_title_area();
 
-    // Fixes misaligned message_view and hidden
-    // floating_recipient_bar.
-    navbar_alerts.resize_app();
+    // Fire our custom event
+    $("#message_feed_container").trigger("message_feed_shown");
 
     // This makes sure user lands on the selected message
     // and not always at the top of the narrow.