--- a/static/js/ui_init.js
+++ b/static/js/ui_init.js
@@ -658,6 +658,8 @@ export function initialize_everything() {
 
     // All overlays must be initialized before hashchange.js
     hashchange.initialize();
+    resize.initialize();
+
     unread_ui.initialize();
     activity.initialize();
     emoji_picker.initialize();