<!-- What's this PR for?  (Just a link to an issue is fine.) -->
We had received a complaint on chat.zulip.org [#feedback > pressing n is slow](https://chat.zulip.org/#narrow/stream/137-feedback/topic/pressing.20n.20is.20slow) about navigation with
the keyboard `n` key being significantly slow (~5 seconds), the first
time `n` was pressed when starting from the recent topics view.

It was difficult to reproduce the amount of lag that was reported, but
running chrome with the profile tab set to 4x slowdown helped get
close to it.

Based on profiling from the original report, as well as locally with
chrome set to 4x slowdown, led to the realization that recent topics
to stream navigation involved a lot of dom thrashing, and so this
series of commits aims to prevent this path from causing forced
reflows.
<details>
<summary>
Profiling data
</summary>

## Profiling data:
### Generation process:
#### one time unreads setup
1) Start server, open chrome page to zulip dev, login as Iago.
2) Send 1 message in # ビデオゲーム.
3) Send 1 message in # 조리법 😎.
4) Logout, login as Cordelia, do not read the unread messages.

#### taking samples
1) Checkout to desired commit.
2) Start server, open chrome new tab, open profiler.
3) Separate the profiler dock to its own window. (This is important to ensure that all profiles are run on the same viewport dimensions.)
4) navigate page to zulip dev, this should log you in as Cordelia and open the recent topics view.
5) Set profiler to 4x slow down. (This is only really necessary if your machine is fast enough that we need to exaggerate the values.)
6) Start recording the profile.
7) Click Support in stream list.
8) Click Recent Topics in top left corner.
9) Repeat (7) and (8) two more times, to have one initial sample and two subsequent samples.

### main:
1) 2.39 s (screenshot)
2) 832.80 ms
3) 815.04 ms

### first commit:
1) 2.57 s (screenshot)
2) 879.65 ms
3) 764.52 ms

### second commit:
1) 2.57 s (screenshot)
2) 787.05 ms
3) 771.08 ms

### third commit:
1) 2.69 s (screenshot)
2) 814.88 ms
3) 746.61 ms

### fourth commit:
1) 970.82 ms (screenshot)
2) 648.25 ms
3) 676.63 ms

Here's a Zip with all of the profiles, which can be loaded into the chrome profiler for inspection:
[Profiling data.zip](https://github.com/zulip/zulip/files/7538738/Profiling.data.zip)

**Profile screenshots:** <!-- If a UI change.  See:
  https://zulip.readthedocs.io/en/latest/tutorials/screenshot-and-gif-software.html
  -->

### main:
![profile-on-main](https://user-images.githubusercontent.com/33805964/141783752-2be8b593-cbfc-4b1b-983a-fa5c7edafb0b.JPG)

### first commit:
![profile-after-first-commit](https://user-images.githubusercontent.com/33805964/141783756-d81a48ec-2740-4571-a075-54631ad0ff36.JPG)

### second commit:
![profile-after-second-commit](https://user-images.githubusercontent.com/33805964/141783759-2c72c9d4-6c8f-4261-8cbb-9a58a5331abb.JPG)

### third commit:
![profile-after-third-commit](https://user-images.githubusercontent.com/33805964/141783747-86557c4a-dd71-4771-9aba-01aaf4ba51cf.JPG)

### fourth commit:
![profile-after-forth-commit](https://user-images.githubusercontent.com/33805964/141783757-96dccc91-cbac-42db-9b79-f502e550a923.JPG)


Finally, testing the "n" hotkey at **6x slowdown**:
before (ie main):
![image](https://user-images.githubusercontent.com/33805964/141784562-6d19e0d0-51f7-4b17-b478-80f7f133ed11.png)
after (ie fourth commit)
![image](https://user-images.githubusercontent.com/33805964/141785188-d833d8a0-08f5-4cc5-a4ef-65d5e034ab79.png)
ie
before: 4.44 s
after: 1.04 s
[profiles at 6x slowdown.zip](https://github.com/zulip/zulip/files/7538752/profiles.at.6x.slowdown.zip)
</details>