@YashRE42 2021-11-15
Do you think we should add code comments before each of the `window.requestAnimationFrame` calls to explain their purpose?