@timabbott 2021-11-15
Thanks for doing this debugging @YashRE42!  A few high-level thoughts:
* At a high level, usually `requestAnimationFrame` used this way is a workaround for a bug that we should fix to eliminate the animation frame (like your first commit).  So we should try to find those bugs :).
* I wonder if the `hide_history_limit_notice` piece could be instead addressed by having that `hide` action happen inside `recent_topics_ui.show()`?  It kinda logically goes with `narrow_state.reset_current_filter` in terms of what state should happen at that time.  Perhaps we want to extract a little function `narrow.reset_ui_state` that calls `message_scroll.hide_top_of_narrow_notices`, `hide_empty_narrow_message`, and `message_scroll.hide_indicators`, and is called from all of `narrow.activate`, `narrow.deactivate`, and `recent_topics_ui` when leaving a view?
* For the `narrow.save_pre_narrow_offset_for_reload` change -- I think the actual bug here may be that we shouldn't be calling that function at all when switching from `recent_topics` to a message feed.  Its purpose is to save the scroll position from your previous view, which I think would need to have happened when switching away from the message feed to recent topics for it to be useful. I'm not sure what the right refactoring is: @amanagr thoughts?  I think structurally we kinda want that function to be called in 3 places: (1) `narrow.deactivate` (I think we'll have to do so in https://github.com/zulip/zulip/issues/20066), (2) in `narrow.activate` when `current_message_list` is interesting, and (3) when displaying recent topics and thus hiding whatever narrow we were visiting previously, before the `$("#message_feed_container").hide();` call.  I think structurally, the call to `narrow_state.reset_current_filter` should be resulting in the a state change that the `narrow.activate` logic can check?  I'm not totally clear on the right refactor, but I think fundamentally the problem here is we needed to do further refactoring to the `narrow.js` state machine when moving "recent topics" to be a central view to not have the assumption that a message list is always visible.  But one piece might be this?
```
diff --git a/static/js/narrow.js b/static/js/narrow.js
index 2a4011b9eb..19a28d4f3e 100644
--- a/static/js/narrow.js
+++ b/static/js/narrow.js
@@ -187,12 +187,15 @@ export function activate(raw_operators, opts) {
          or rerendering due to server-side changes.
     */
 
+    const start_time = new Date();
+    const was_narrowed_already = narrow_state.active();
+
     if (recent_topics_util.is_visible()) {
         recent_topics_ui.hide();
+    } else {
+        save_pre_narrow_offset_for_reload();
     }
 
-    const start_time = new Date();
-    const was_narrowed_already = narrow_state.active();
     // most users aren't going to send a bunch of a out-of-narrow messages
     // and expect to visit a list of narrows, so let's get these out of the way.
     notifications.clear_compose_notifications();

```