@amanagr 2021-11-16
Yeah, I think your suggested change will have the same effect as `requestAnimationFrame` on the last commit. Maybe we don't need to refactor more than that.