@timabbott 2021-11-16
I think my change is incomplete; we probably need to call `save_pre_narrow_offset_for_reload` when entering recent topics as I described, as well, since that's the point where we may discard your old scroll position.