@YashRE42 2021-11-26
I don't really understand the purpose of calling `save_pre_narrow_offset_for_reload` when entering recent topics, since the only place where `pre_narrow_offset` is in `preserve_state` in `reload.js`... and if we're currently on recent_topics when we reload.. doesn't the offset not matter?