@YashRE42 2021-11-26
Aside from the above, I've made the requested changes. Commit 49c234a659ee06c509bb72f2dbd4e57ea2612774 (recent_topics_ui: Delay resize_app to next animation frame via rAF cb.) still uses a `requestAnimationFrame` for `navbar_alerts.resize_app`... and a way to eliminate that rAF would be do address the fact that we probably shouldn't be calling that unconditionally? Solving that would probably involve adding some basic state to `navbar_alerts` and exposing a `maybe_resize_app`... but maybe we want to keep the rAF for the case where there actually is an alert banner? 


## Profiling data:
Generated through same process as before:

### main:
1) 2.26 s (screenshot)
2) 572.56 ms
3) 607.77 ms

### 1a002a12ac0cad6c435ce3f6c9aa6061bce50209 (recent_topics: Call show() on message list underpadding b/f container.):
1) 2.20 s (screenshot)
2) 643.80 ms
3) 613.80 ms

### 49c234a659ee06c509bb72f2dbd4e57ea2612774 (recent_topics_ui: Delay resize_app to next animation frame via rAF cb.):
1) 2.25 s (screenshot)
2) 564.88 ms
3) 569.25 ms

### d99cf811f2d7092cae858550e6741103a97b41f9 (recent_topics: Move narrow.reset_ui_state() calls up into .hide().):
1) 2.26 s (screenshot)
2) 565.39 ms
3) 598.35 ms

### bf2cf9e7c4fdc2107a48b7d8783e72a0f2ef8ac1 (narrow: Don't call save_pre_narrow_offset_for... when coming from rt.):
1) 726.91 ms (screenshot)
2) 501.22 ms
3) 522.58 ms

Here's a Zip with all of the profiles, which can be loaded into the chrome profiler for inspection:
[Profiling Data.zip](https://github.com/zulip/zulip/files/7610253/Profiling.Data.zip)

## Profile screenshots:
### main:
![](https://user-images.githubusercontent.com/33805964/143618861-fd301db0-a970-4107-9ede-378ac7a45b0a.JPG)

### 1a002a12ac0cad6c435ce3f6c9aa6061bce50209 (recent_topics: Call show() on message list underpadding b/f container.):
![](https://user-images.githubusercontent.com/33805964/143618919-24f44fc8-3501-4be0-9167-9cb7340a18fa.JPG)

### 49c234a659ee06c509bb72f2dbd4e57ea2612774 (recent_topics_ui: Delay resize_app to next animation frame via rAF cb.):
![](https://user-images.githubusercontent.com/33805964/143618939-b70f6fdf-8e8d-4940-b07d-141cde0c86a6.JPG)

### d99cf811f2d7092cae858550e6741103a97b41f9 (recent_topics: Move narrow.reset_ui_state() calls up into .hide().):
![](https://user-images.githubusercontent.com/33805964/143618951-3134fa31-99e5-4b87-9bdb-004b23b83230.JPG)

### bf2cf9e7c4fdc2107a48b7d8783e72a0f2ef8ac1 (narrow: Don't call save_pre_narrow_offset_for... when coming from rt.):
![](https://user-images.githubusercontent.com/33805964/143618970-f912fb1f-3cbc-4fe8-89b1-a8494f817d94.JPG)