@timabbott 2021-11-28
Could we just move the call to `reset_ui_state` to sit above the `recent_topics.is_visible` check, rather than moving it to be called within `recent_topics_ui.hide`?  That feels a bit like an abstraction violation.  I think the only different would be whether calling `recent_topics_ui.hide` from `hashchange.js` ends up calling `reset_ui_state` via calling `recent_topics_ui.hide` -- and I'm not sure what's correct in that code path.