@timabbott 2021-11-28
> I don't really understand the purpose of calling `save_pre_narrow_offset_for_reload` when entering recent topics, since the only place where `pre_narrow_offset` is used in `preserve_state` in `reload.js`... and if we're currently on recent_topics when we reload.. doesn't the offset not matter?
> ... I haven't actually checked, I just expect that if a reload is forced while we're on recent topics, we will stay on recent topics.

That is correct, but we have plans to make the `Back` button preserve scroll position, and those will involve wanting the pre-narrow offset to be saved whenever we leave a view that currently has a scroll position.